Virtual SmartZone
I'm back in Education baybay, so that means more powershell, this one is to interact and login to a virtual smartzone. Not on-prem. It's tested using 5.1.0.0.496. So far I have figured out logging in, I am working on emulating some other basic API calls that I require in my job, any other features, open an issue.